refactor: abstract away report filters
Abstracts away the report filters so they can be reused in other views. Also deletes some unused styles.
