+++ report-app/src/app/pages/report-list/report-list.html
@@ -3,38 +3,7 @@
     <message-spinner message="Loading reports"/>
   } @else {
     <div class="toolbar">
-      <div class="filter-container">
-        @if (allFrameworks().length > 1) {
-          <select (change)="selectedFramework.set($any($event.target).value)">
-            <option value="">All Frameworks</option>
-            @for (framework of allFrameworks(); track framework.id) {
-              <option [value]="framework.id">{{ framework.displayName }}</option>
-            }
-          </select>
-        }
-        @if (allModels().length > 1) {
-          <select (change)="selectedModel.set($any($event.target).value)">
-            <option value="">All Models</option>
-            @for (model of allModels(); track model.id) {
-              <option [value]="model.id">{{ model.displayName }}</option>
-            }
-          </select>
-        }
-        @if (allRunners().length > 1) {
-          <select (change)="selectedRunner.set($any($event.target).value)">
-            <option value="">All Runners</option>
-            @for (runner of allRunners(); track runner.id) {
-              <option [value]="runner.id">{{ runner.displayName }}</option>
-            }
-          </select>
-        }
-        @if (allLabels().length > 0) {
-          <multi-select
-            [options]="allLabels()"
-            [(selected)]="selectedLabels"
-            [label]="selectedLabels().length === 0 ? 'Filter by labels' : `Showing reports with ${selectedLabels().length} label(s)`"/>
-        }
-      </div>
+      <report-filters #filters/>
       <div class="button-group">
         @if (isCompareMode()) {
           <button class="outlined-button" (click)="toggleCompareMode()">Cancel</button>
@@ -48,7 +17,7 @@
       </div>
     </div>
 
-    @for (group of reportGroups(); track group.id) {
+    @for (group of filters.filteredGroups(); track group.id) {
       <div class="report-item">
         <a [routerLink]="['/report', group.id]">
           <div class="report-score-container">
