+++ report-app/src/app/shared/report-filters/report-filters.scss
@@ -1,0 +1,9 @@
+:host {
+  display: flex;
+  gap: 1rem;
+}
+
+select,
+multi-select {
+  max-width: 200px;
+}
